services/api/pyproject.toml
[project]
name = "xiaolou-api"
version = "0.1.0"
description = "Production Python API for XiaoLouAI"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13,<2.0",
  "asyncpg>=0.29,<1.0",
  "celery[redis]>=5.4,<6.0",
  "cryptography>=42,<47",
  "fastapi>=0.115,<1.0",
  "httpx>=0.27,<1.0",
  "orjson>=3.10,<4.0",
  "prometheus-client>=0.20,<1.0",
  "pydantic-settings>=2.5,<3.0",
  "python-multipart>=0.0.9,<1.0",
  "redis>=5.0,<8.0",
  "sqlalchemy[asyncio]>=2.0,<3.0",
  "structlog>=24.0,<26.0",
  "uvicorn[standard]>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<10.0",
  "pytest-asyncio>=0.23,<2.0",
  "ruff>=0.6,<1.0",
]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
